Add preferences interface and plist read/write
Initial integration of preferences/plist code. The first-run interface
will trigger if there is no existing plist file. It will allow modifying
the URL log file location and the text editing application. A
preferences option still needs to be added to each list dialog, but that
will just trigger the same interface that the first-run dialog uses to
modify the preferences.

The experimental environment variable configuration is removed. It was
never really meant for public release.