Add file viewer preference and UI for changing it
Still need to add an "Open Log File" menu item to all of the menus that
have an "Edit Log File" item.

Also deleted all of the duplicate go_back() methods in most of the
controller classes and moved it to the base controller class that most
of the other controllers subclass.