Fix error handling for missing browser data
Some errors such as a missing URL weren't handled, resulting in a script
error. Those errors are now caught and handled with an error dialog.

Firefox doesn't throw an error for a missing URL because doing a GUI
copy when nothing is selected doesn't actually copy anything and doesn't
result in an error because the old clipboard data is still there.