Fix Yosemite 10.10.5 AppleScript open file bug
An error occurred with a call such as:

    open alias (file "mac:path:to:file")

The fix was to coerce the file path to a string first so that the call
ended up being:

    open alias "mac:path:to:file"

First noticed the bug after Yosemite Security Update 2016-001, but not
sure exactly when the OS X AppleScript change occured. I don't see
anything in those release notes or earlier release notes about any
AppleScript changes like that. Maybe it was a side effect of some
unrelated system change. Or maybe it was just a stealth fix for a
previously undocumented bug.